Preserve const refinements
Summary: Consts are read-only, so their refined types can be retained during a havoc.

This means for example that Flow now understands it's safe to close over type-refined consts in a local function that escapes, e.g.:

  const n: ?number = ...;

  if (n) {
    // this is safe: if n is truthy here, it will be truthy whenever the lambda runs
    f(() => { var x:number = n; });
  }

The implementation is extremely straightforward - just as with read-only internal vars like `.this` and `.return`, we simply make havoc a no-op for const entries in scope tables.

Reviewed By: gabelevi

Differential Revision: D2553608

fb-gh-sync-id: e89c034fe19256924218d917bcba5d168f140083