refactor docblock parsing
Summary: Refactors docblock parsing so that it can happen before parsing the file.

In a followup, I will rely on this to control the `types` parsing mode depending on whether the docblock includes `flow` (so, you can see why we can't use the comments extracted during parsing to look for `flow` anymore).

Two notable things:

- still uses the lexer to extract the first comment
- collects all of the directives that we care about (`flow` and `providesModule`, plus `preventMunge` which we can use in a future diff... or I can remove it until we care) at once. previously, we re-parsed the docblock in several places.

Reviewed By: gabelevi

Differential Revision: D2558771

fb-gh-sync-id: 2627587c7681e554fd513792b9f1b13ad6589f31