remove debug from modes
Summary: easy enough to thread around. this also changes `add_output`'s extra logging from `--debug` to `--verbose`, because it makes the most sense in the context of the rest of `--verbose`'s output to know when an error happened, and I don't really want to store `debug` in the context to make it accessible from `add_output`. :)

Reviewed By: bhosmer

Differential Revision: D2569149

fb-gh-sync-id: 16efe9d803350b39af81d067058eb6ea4fe58ada