Look for @flow in first comment, rather than first token.
Summary: D2558771 changed the way that Flow looks for `flow` comments. Previously, it
would match `flow` anywhere in the program text (it didn't even need to be in a
comment). That diff changed it so the first token had to be a comment.

Unfortunately, Nuclide needs `'use babel';` to be the first token, since Atom
looks for that to figure out when to do transpilation (that is a whole other
area for discussion but that is the way things are at this moment).

This diff relaxes the requirement from "first token" to "first comment."

Reviewed By: jeffmo

Differential Revision: D2575194

fb-gh-sync-id: 1dc56f31fde3e3d1e2c19feca4cb21e0b0efff45