Fix an edge case in the docblock parser
Summary: In the prior diff we changed the docblock parser to look for the first comment (rather than comments before the first token). However, it didn't consider the edge-case where a file might consist only of a comment and nothing else. Normally this doesn't matter, but for the Haste module system, a `providesModule` in a comment defines the module in the system -- so it's actually meaningful.

This just changes the parsing to scan for comments until we hit a token other than `T_STRING` or `T_SEMICOLON` (the only two tokens we need to work around for now)

Reviewed By: nmote

Differential Revision: D2575333

fb-gh-sync-id: c992390259400e63920ec56918333cf86a84a81c