[hack] Make Call-time Pass-by-reference a Parse error
Summary: Call-time pass-by-reference is when you use the reference operator & on the
input of a function parameter. For example:

```
$arr = array(1, 2, 3);
array_shift(&$arr);
```

This was deprecated in PHP 5.3 and causes a fatal in 5.4. In PHP 7 this is a
parser error according to https://3v4l.org/W6oDd This is a fatal in HHVM as well.

This diff makes call-time pass-by-reference a parser error in Hack as well

Reviewed By: achow, dabek

Differential Revision: D2566300

fb-gh-sync-id: d3cafea1c15953a05cb479dfd4f7c927a6560c47