add --root arg to most client commands
Summary: adds a `--root=<path>` arg to client commands that take filenames (so, all of them except `start`, `status`, `stop` and `check`) so that you can ensure the correct `.flowconfig` is used when inside a directory with multiple `.flowconfig` ancestors.

in theory this could also be a minor perf gain since flow won't need to walk up the the file tree on every command, but that seems exceptionally minor.

Reviewed By: gabelevi

Differential Revision: D2561803

fb-gh-sync-id: 7366caa284acdb0a6f681835e65a3c12df143075