[hack] Print out hh.conf on startup
Summary: Since it can change after we start, we should print it out in our logs
to have a record of what settings the running hh_server is using.

I also snuck in a change to stop printing No_loader when saved state v2 is
disabled -- if we're encouraging people to read the logs, we should avoid
putting in misleading messages that look like errors when they are actually
benign.

Reviewed By: dabek

Differential Revision: D2570040

fb-gh-sync-id: 65074921d1211067d6ce31c0bc53d91e6862b7e2