add @noflow annotation
Summary: now that we refuse to parse types in non-flow files, we need an escape hatch to temporarily disable flow without removing all the types (a fairly common debugging strategy).

this diff adds a `noflow` annotation. I went with that instead of `flow ignore` because it's easier to grep for `flow` and `noflow` than to grep for `flow(?! ignore)` and `flow ignore` when you want to locate all of the flow and no-flow files (e.g. we do this to loosely track code coverage).

Fixes https://github.com/facebook/flow/issues/978

Reviewed By: bhosmer

Differential Revision: D2578582

fb-gh-sync-id: 31f64dbc3b6b7c968cc0a8166fe0d8d0a7cd9a10