fix package not found error
Summary: We were failing to pick up some paths in our symlink spidering due to a
reappearance of the always-entertaining string-prefix-is-not-path-prefix
bug, e.g. `/a/b/c` is not a prefix of `/a/b/cdef`.

This fixes the "package.json not found" errors we were seeing in certain
Nuclide projects (also edited that error message to clarify that it was
an internal heap where they weren't getting found, not the file system).
It may also fix other things, as the bug is in a function that's  used
elsewhere.

Reviewed By: mroch

Differential Revision: D2579376

fb-gh-sync-id: fcdec8c64bf07dd25e3c8c39d0546a0b35873352