Allow callbacks passed to Promise functions to return whatever
Summary: This makes it easier to use arrow functions with Promises since you
can write

  new Promise((resolve, reject) => resolve(1));

instead of having to write

  new Promise((resolve, reject) => { resolve(1); });

Note: we only need to relax the return types of callbacks that the user of the
API is passing in. Callbacks that are passed TO the user can still use `void`
as a return type.

Reviewed By: jeffmo

Differential Revision: D2559819

fb-gh-sync-id: 86d708edafe405a25edfd9b00a388bdeab30677d