Detect function param name clashes
This commits adds a new kind of implicit let binding for function
parameters. This helps us catch two errors: 1) param bindings that
shadow another param binding and 2) let bindings that shadow a param
binding.

It's actually important that we don't create an intermediate lex scope
for params to make this work, otherwise we wouldn't be able to catch
error condition 2 (above).

Tests added in previous commits ensure that we don't behave badly due to
the lack of a separate param scope. (This is because default param exprs
are "evaluated" before the function body, and are thus unable to ref
bindings from the function body.)