[hack] Treat values of heterogeneous arrays downcasted from AKshape as Tany if any of them was Tany
Summary: This is mirroring diffusion/FBCODE/browse/master/hphp/hack/src/typing/typing.ml;8008a8cf243001065f24d0fe5282d56008843d3e$924-934 logic, in the case when we went through AKshape first. See unit test for explanation.

I plan to disable the one in typing.ml too, but put it behind config setting - for people who are just doing the migration and might want to be less strict and aggressive, we want to keep it, and have it work even with AKshape.

Reviewed By: dlreeves

Differential Revision: D2588341

fb-gh-sync-id: 691bac28a6408ddc5785d4a2b07873762e356e92