add coverage command
Summary: adds `flow coverage path/to/file.js` which prints the percentage of ranges (basically expressions, I think) that have a type other than `AnyT` and `UndefT` (`UndefT` includes things that we know but don't know how to print, so we should drive that down to make this accurate).

```
Covered: 79.68% (251 of 315 statements)
```

also has a `--color` flag which highlights which parts of the file are missing types (red).

Reviewed By: gabelevi

Differential Revision: D2345020

fb-gh-sync-id: 4a18f2c910aec2d0cc98edb9d1938f9b3f3a8ed8