Connection-fixes
This fixes a few things we discovered on OSX.

First on my laptop, `/tmp` is a symlink to `/private/tmp`. Some commands would
get the `realpath` for the tmp dir and some wouldn't so you would end up with
some commands using `/tmp/flow` for the socket path and some using
`/private/tmp/flow`. For socket paths that are less than 104 characters using
`/tmp` and more than 104 characters using `/private/tmp`, you would end up with
clients that can't talk to servers, since the paths > 104 characters would be
shortened automatically.

Second, we set a limit on the number of pending socket connections. During
initialization we were hitting this limit on OSX but not on Linux. It looks like
`Unix.shutdown_connection` and `close_in_noerr` don't actually remove the
connection from the pending connection count on OSX (at least if the server
isn't listening). To fix, we can just memoize the connection during
initialization.