Fix connect race condition
Summary: I noticed a race condition around `flow start` and how other commands,
like `flow status`, use it to start a server. This diff fixes the behavior of
running N commands in parallel that all try to start a server at the same time.

If you run `flow start` N times in parallel, one of those commands will start
the server and the others will fail to get the lock. However the failing
`flow start` commands will fail as late as possible since all the early lock
checks will pass. This diff recognizes when that happens and sets the exit code
to properly reflect that `flow start` failed to get the lock, indicating that
there's already a server running.

Building on that, the client connect code now recognizes when a `flow start`
fails due to an existing server and treats that as a successful start, as long
as the server is there after 1 second.

Reviewed By: mroch

Differential Revision: D2592751

fb-gh-sync-id: 83305992466197cfd82df37a63720e2ae37aff42