fix merging refinements that don't exist in parent scope
Summary: if a refinement didn't exist in the base scope but did exist in both child scopes, `merge_env` was dropping the refinement.

this comes up in `if` statements, where the base scope is before the statement and has no refinement, `child1` contains the refinement from the if's test (e.g. `x.foo != null`), and `child2` contains the negation of that refinement (e.g. `x.foo == null`). normally it's fine to drop the refinement since `(x.foo != null) + (x.foo == null) == x.foo`, but if one or both of the scopes has a definite assignment then dropping the refinement blew away the assignment's effects.

Reviewed By: bhosmer

Differential Revision: D2606686

fb-gh-sync-id: a1dc9c5a3bb8f041f6269aaf6f572bad4d4f5d1d