extract ops into their own JSON field
Summary: instead of prepending the operation to the head of an error's list of messages, this diff carries it around as a separate field (like traces). it is prepended when printing to the console, as before, but is printed as its own field in the JSON output.

this is one step closer to regularizing the JSON output, so that they're always "source\nmessage\ndest", not sometimes "op\nError:\nsource\nmessage\dest" (and really, we should have separate fields for "source", "message" and "dest" instead of a list of reasons at all...)

Reviewed By: bhosmer

Differential Revision: D2583373

fb-gh-sync-id: d8e060db552456344df1a543ab7d78af7fc4e331