fix refinement of if statements that contain loops
Summary: for loops and while loops (possibly among other things) return with a different scope than they started with. if statements merge the then and else scopes into the original scope at the end of the if statement, but they merge the then and else scopes as they were at the beginning of the if statement. this diff makes them use the scope that's actually current at the end of each branch, making the branches immune to anything that modifies the scope.

Reviewed By: bhosmer

Differential Revision: D2609813

fb-gh-sync-id: 0c25ba5d17ad23258a5a3edb6b35768a74c832cf