Default param values can reference earlier params
Summary: Defer evaluation of default expr until mk_body

This commit introduces a module, `Params`, which encapsulates lots of
information that was passed around in a tuple of interdependent values
between the various functions that build up function types.

The original motivation for this refactor was to defer the evaluation of
the default expressions. Adding yet another value to the "params_ret"
tuple was a step in the wrong direction, so I tried something else.

This also fixes a case where users could force an unexpected BoundT
error, by adding a default expr to a generically-typed param. Now the
default expr is evaluated within the generate_test callback.

This commit is larger and noisier than I wanted it to be, which is a
consequence of the added module. The module depends on some functions,
and other functions depend on the module, but all of those functions
were simultaneously declared with `and`. In order to define the module,
I needed to split the declarations, which involved mov
Closes https://github.com/facebook/flow/pull/872

Reviewed By: bhosmer

Differential Revision: D2595701

Pulled By: jeffmo

fb-gh-sync-id: 790a0d16f845d5326536228711ca27b884310b2a