[hack] Consider AKempty to be fully checked in coverage metric
Summary: AKempty inherited this coverage_level from AKany, but it's not necessary. There is nothing unknown about the type of AKempty - it's the type of an empty array, and we correctly update it to the more specific type on first lvalue usage. In a fully typed, strict code (no untyped functions, no functions modifying the type by reference) a variable typed AKempty is always an empty array.

Reviewed By: int3

Differential Revision: D2613536

fb-gh-sync-id: 369c41a7b7cd70046383b40d7157220d59439143