[easy] fix lib dev vs unchecked module precedence
Summary: When an unchecked (non-@flow) module and a library module declaration share a name,
the intended behavior is that requires and imports using that name resolve to the
library definition. This was partially broken: if the require/import used anything
other than the exact module name (e.g., a path), the library definition would be
missed and the unchecked module resolved instead. This diff fixes.

Added tests to check desired behavior in haste and node.

Reviewed By: avikchaudhuri

Differential Revision: D2615141

fb-gh-sync-id: 1942e0d72ccdd9512aaceb1c873c488579bcda36