Remove support for `import type * as`
Summary: This syntax started life with a version of CommonJS<->ES interop that we wound up not going with. We left these semantics laying around for a while to leave time to phase things out, but it's been months at this point so it's time to kill it off.

(Specifically, at the time, we were going to map `import * as Foo from "Foo";` to mean the ~same thing as `var Foo = require("Foo")`; But then we decided that it was better to map the `module.exports` value to the default export in a CJS module rather than assume it's an object that can be directly mapped to a ModuleNamespaceObject -- which is the *only* thing `import * as` can yield as per the ES spec)

So, since `import * as` can only ever be a value (and can never be a value that possibly also represents a type -- like a class), `import type * as` couldn't have any sensical semantic meaning other than be an error. In theory `import typeof * as` might make sense (to extract the type of the ModuleNamespaceObject) -- but given that we don't have it now and haven't had any asks for it so far, I figure we can defer implementing that until we have a real reason to.

Reviewed By: avikchaudhuri, gabelevi

Differential Revision: D2621526

fb-gh-sync-id: 9c1a9ce9a2e2a8f1ea38f8133bb93f6656a96a09