Class field initializers
Summary: Initial support for `esproposal.class_{instance,static}_fields=enable`.

Tests describe most of the expected semantics, but the proposal is being discussed and iterated on here: https://github.com/jeffmo/es-class-properties

Note that this is a very early stage proposal and so there is some reasonable chance this will change as it moves forward.

Reviewed By: gabelevi

Differential Revision: D2588595

fb-gh-sync-id: cfa93b9081593470b205d21d6e3c82b4f78a981d