autocomplete coverage boost
Summary: This diff is based on a conjecture that a lot of logged autocomplete "failures"
are due to inadequate type coverage.

Instead of failing, we now succeed with an empty map when we don't have enough
type information. Also, autocomplete involves a merging operation that was
unnecessarily strict when we didn't have enough type information about a type
variable, so this diff relaxes that operation.

Reviewed By: mroch

Differential Revision: D2614215

fb-gh-sync-id: 1e76218e4d1b2b81106b5a4509678240c7466d0a