library definition overrides
Summary: Flow builds its set of globally-available builtins from multiple
library files. The core set is provided by flow itself, and additional
library paths may be specified in a project .flowconfig and/or on
the command line.

Previously, if a given name was bound to definitions in multiple
files, the resulting builtin's type would be the union of the
encountered definition types, exactly as if the same property
on the global object had been set more than once.

This diff implements library overrides, so that a definition
found in a later file replaces a same-named definition found
in an earlier file, according to the following specification order:
```
1. the Flow core library files
2. the files found in the paths specified in the .flowconfig [libs] (in listing order)
3. any paths supplied on the command line via --lib
```
Importantly, when one definition overrides another, the later
one replaces the earlier one completely, even in references
from earlier library files. This makes it possible to override
a definition completely without needing to create overrides
for every definition that refers to it.

Reviewed By: gabelevi

Differential Revision: D2599884

fb-gh-sync-id: 1b370d554d913ce32bbc9c68d4be85b61ca3422f