[hack] Don't return Tany if we indexed an array containing a Tunresolved
Summary: Part of the reason why arrays are typed so poorly is because we give up on
tracking them so easily. For instance if we have a map like array with mixed
keys we will return Tany when you index into it. This is really surprising
behavior and leads to lots of bugs especially in strict mode.

I think I understand the rational, since arrays are used in a ton of different
ways just give up instead of reporting an error. With HH_FIXME I think its
better to report an error and have the user explicitly ignore the error.

If this is seen as too painful there are a few options:

Neither are my preference but curious what others think.

Reviewed By: int3, jwatzman

Differential Revision: D2281048

fb-gh-sync-id: 13fee133ff5ac796241596d6d4a071bfeafaf504