[cleanup] move Abnormal and Refinement to their own files
Summary: I know this exposes things that were technically private in `Type_inference_js` before, but that file is too big so splitting things up is going to be necessary anyway, I think.

We can possibly use build dependencies (like, put this stuff in a folder that we only add to the search path while building `type_inference_js.ml`) to make this stuff private again, but I'm not worried enough about it to do that.

The number of files in `src/typing/` is also getting out of hand, but i think it'll get worse before it gets better (once we can separate dependencies well, I can move them into subfolders).

Reviewed By: bhosmer

Differential Revision: D2635673

fb-gh-sync-id: 6a183196b7726f93b0c2676552ca0f46c685f9c7