Speed up runtests.sh by running tests in parallel
Summary: For me this speeds up runtests.sh from around 120s to around 16s.

It basically runs 16 tests in parallel and reaps the test results in order.
Whenver it reaps a test it starts a new one. So a slow test can lead to some
wasted CPU cycles (since we'll wait to reap it), but most of our tests are
really fast so this saves lots of time.

Reviewed By: mroch

Differential Revision: D2638086

fb-gh-sync-id: 075bbabed04d52a5156a3233e4af9d81d89fb763