fix position of inherited properties
Summary: if you call `get_prop` on a prop that exists in the given property map, then we look it up and reposition it. if you call `get_prop` on a prop that exists on `super`'s (or above) property map, then we call `lookup_prop` and do not reposition it, resulting in using the definition's position instead of the position of the lookup.

this is inconsistent and leads to bad "errors at a distance". see the tests for some examples; note how all of the errors are much more local now, as they would be without inheritance.

most notably, the new test in `classes/B.js` previously errored across file boundaries, saying that `A.js: module 'prop'` was incompatible with `B.js: number`. so the error would even show up on the wrong file before!

Reviewed By: bhosmer

Differential Revision: D2641265

fb-gh-sync-id: 9d2f00de0a1c127a8c6bf617163cf95a43e80891