[hack] Finish converting asio to HNI
Summary: Cheat the standard HNI-NativeData model by leaving the extra data
for asio wait handles at the end of the ObjectData struct.
Normally this would be verboten, but Asio objects have the
unique property of being effectively final (private ctor)
and can only be constructed by internal APIs.

This conversion takes advantage of this and allows the
object structure to remain as-is, greatly simplifying
this conversion from what it would otherwise have been.

Reviewed By: jano

Differential Revision: D2476884

fb-gh-sync-id: 861b41b1978b32b1cafc329250546fc24b735292