Preserve tuple information in Function.prototype.apply
Before this change, the overall element type of the array was passed
into all arguments of the target function. This made it impossible to
use apply for functions that took arguments of multiple parameters.

The added ApplyT type holds on to the target FunT until the arguments
array is resolved, and then calls the function. Because we wait for the
arguments array to be resolved, we can pull out the tuple types to use
as parameter types.

Fixes #64