Preserve tuple information in Function.prototype.apply
Summary: Before this change, the overall element type of the array was passed
into all arguments of the target function. This made it impossible to
use apply for functions that took arguments of multiple parameters.

The added ApplyT type holds on to the target FunT until the arguments
array is resolved, and then calls the function. Because we wait for the
arguments array to be resolved, we can pull out the tuple types to use
as parameter types.

Fixes #64
Closes https://github.com/facebook/flow/pull/1104

Reviewed By: svcscm

Differential Revision: D2657396

Pulled By: mroch

fb-gh-sync-id: 861f7eac56e9afdb9fa58d0e4578a5b462b36ae1