suppress ops when doing property lookups
Summary: we do a good job with positions on property lookups. `reason_op` is always the position of the lookup (e.g. `foo.bar`) and `reason_prop` is always the position of the prop (`bar`), and errors always point at the prop (`property 'bar': Property not found in...`), so the op doesn't provide any context.

this suppresses ops while doing property checks. I chose to clear/restore the stack instead of pushing something that we then suppress, since it avoids implicit coupling whether we want ot show an op via the op's position, error position, and suppression heuristics.

Reviewed By: bhosmer

Differential Revision: D2653667

fb-gh-sync-id: 0a0c045e6d5713a69f904e2462d660e9407c92b0