Refining rules for Tainted types. Calls to Object properties emit an error.
Summary: Before the changes, it was possible to do the following without flow reporting any errors:
function foo(x : $Tainted<string>, o : Object) {
  o.f(x);
}

Objects are not strictly type-checked with respect to operations, but this is a source of false negatives in the context of taint analysis, since the object could then be passed along to an annotated function expecting a non-tainted value. This example is now included in the tests.

Reviewed By: avikchaudhuri

Differential Revision: D2573859

fb-gh-sync-id: 5c0b9e1871d75b471aef4b9bfda2e5f2459916d3