add function.name, function.length
Summary: properties on functions should be looked up on `Function` from the core lib, but weren't because all `FunT`s had `MixedT` as their root type. this diff adds `FunProtoT`, which serves as the root type for functions, like `MixedT` is the root type for objects (hey, let's add an `ObjProtoT` and get rid of the confusing multiple uses of `MixedT`!).

then I added the missing props from `Function.prototype`, `name` and `length`.

Reviewed By: bhosmer

Differential Revision: D2666865

fb-gh-sync-id: 38909df0cf1887a7f014a90f74396ed91d2b0877