default props fix
Summary: ES6 React classes needed to specify a type for static defaultProps for
props checking to work. This requirement wasn't very obvious, so errors were
inadvertently missed.

This diff introduces the concept of an abstract field, specified by using the
type constructor $Abstract on the type of a field. If an abstract field is not
declared in a subclass, it is considered undefined.

We mark static defaultProps of React.Component as abstract. This makes that
field be undefined for any React class that doesn't declare it.

The technical reason why props checking wasn't being triggered when static
defaultProps was missing was that the operation of diffing Props and
DefaultProps, to get the set of required attributes to be passed to React
elements, would remain blocked. This operation is now unblocked. Finally, we
slightly relax the typing rule for ObjRestT, which diffing depends on, to work
for undefined (just like ObjAssignT).

TODO: In future diffs, the concept of an abstract field obviously will be
fleshed out. For example, instead of using a type constructor, we should specify
an abstract field to use a qualifier. Abstract methods should also be supported.

Fixes #398

Reviewed By: gabelevi

Differential Revision: D2611505

fb-gh-sync-id: cf59380bbff34487bfdf5c2ab1dcf9376304739e