Fix module.name_mapper PROJECT_ROOT substitution
Summary: I think the code intended to use `Filename.quote` rather than
`Str.quote`. `Str.quote` escapes things for use in regular expressions. This
means some paths were being escaped in different ways. Someone on Arch Linux
reported that this test would fail when Flow was installed in
`~/Downloads/flow-0.18.1/` but not when it was moved to `~/Downloads/flow18/`.
I think this is because `Str.quote` would escape the `.`'s.

Reviewed By: jeffmo

Differential Revision: D2670282

fb-gh-sync-id: 5b9dabff145d3dd7fa0351994e564d7e235ede7e