fix reason of DiffT
Summary: `reason_of_t` on a `$Diff<t1, t2>` returned the reason of `t1` with no indication that it was a diff. AFAICT, this was only ever hit in `--verbose`, so no big deal.

the `DiffT` handler in flow_js used the reason of `t2`, which caused weird error messages like saying that `property 'name'` was not found in `DefaultProps`, when really it wasn't found in `Props`.

this diff goes further, and makes the reason say `diff of <t1> and <t2>`, and then uses that in the `DiffT` handler as well, so it's consistent.

Reviewed By: avikchaudhuri

Differential Revision: D2648611

fb-gh-sync-id: a16152a64564628fdcdebf36313535bdfa8a5c7e