[easy] rename internal type that represents empty set
Summary: The role of UndefT in the type system is to represent the empty type. It got its
current name because in the beginning we thought the type of undefined should be
the empty type. That has clearly changed, so I'm finally renaming this type to
reflect what it is.

Hopefully this will prevent us from ever using UndefT instead of VoidT to type
something we know could be undefined.

Reviewed By: bhosmer

Differential Revision: D2671230

fb-gh-sync-id: abc921072712d76bc00b2544e192eac5b3125dac