Fix client connect for out of date servers
Summary: Currently we memoize the connection to the server, since there is a
limit to how many pending connections a server may receive. However sometimes
we connect to a server, see that it's out of date, kill it, and start a new
server. This means we need to throw away the old connection and open a new one.

Reviewed By: mroch

Differential Revision: D2676773

fb-gh-sync-id: 6fcf86706ce8c4d6d2784088dd6f93b0a5766703