package incr fix
Summary: When module resolution depends on global packages and the packages are modified, the
recorded stuff gets stale. Unfortunately, since the files that contained such
requires may not have been modified themselves, the stale stuff is not cleared
when rechecking, causing inconsistencies.

For now, we restart whenever packages are modified. We could try something more refined in the future.

Reviewed By: gabelevi

Differential Revision: D2452341

fb-gh-sync-id: a1acea9fa063076fc3797518bd13f59db6d62c8d