[hack] Consider Tunresolved[Tany] to be not checked in coverage metric
Summary: First result of looking at data from {D2675630} and {D2677275} is that a bunch of things that we count as partial look like they should actually be unchecked.

Very often a local variable with value coming straight from an untyped function will be considered partially checked, instead of unchecked. This is because we only check for Tany, while we often wrap stuff in Tunresolved (in particular, the conditional branches integration algorithm does it). This shouldn't change the coverage level of a type.

Reviewed By: int3

Differential Revision: D2682442

fb-gh-sync-id: dc9571b26129b927a5aa41b050831116d8849563