[easy] Bandaid for havoc_call_env exception
Summary: This diff adds a guard around calls in havoc_call_env that were causing
exceptions.

havoc_call_env havocs write-set variables when a function is called.
When it does, it updates the ambient changeset, on the assumption that
the process is taking place during AST traversal of the call site.

This assumption turns out not to always hold: sometimes, the concretization
of a tvar will induce the flow of a function to a call site at merge time, at which
point the ambient changeset is uninitialized.

This diff doesn't address the semantic implications of havocs that don't
occur until merge time - it only prevents crashes. Figuring out the problem
revealed deeper issues with our current approach to write-set refinements,
which will need a different kind of fix - see T9216778 for details.

Reviewed By: gabelevi

Differential Revision: D2695525

fb-gh-sync-id: c8158deea8fee48e166ed7959eb34e4d18f8c2d3