Special connection handling of rechecking state
Summary: At the moment, when the server is rechecking it goes unresponsive.
`flow status` will try to connect a few times (3 by default) and then give up.
So if the recheck takes longer than 3 seconds `flow status` will fail.

This diff makes the connection code aware of the idea of rechecking, and treats
it similarly to initialization. Now `flow status` and other client commands
will just wait for the rechecking to finish and will display the current
status.

Reviewed By: mroch

Differential Revision: D2693983

fb-gh-sync-id: 52e30c46c1a1d81c7fe601e9ca7a74b2e6af100d