Don't print spinner if we are not attached to tty
Summary: If the caller doesn't take care to read from stderr / pipe stderr to
/dev/null, the stderr buffer can fill up and cause hh_client to block
indefinitely. Moreover, if hh_server is waiting on hh_client to send a
message, this can result in the server being unresponsive; that is being fixed
in D2706540.

I also renamed Tty.print to Tty.cprint to reduce potential confusion.

Reviewed By: achow

Differential Revision: D2706714

fb-gh-sync-id: 6224a96fd050396d6bfc135555718a8ffae1efa6