Decompose AnyObjT into AnyT
Summary:
mirrors what `ObjT _, X` does, for each `AnyObjT _, X`.

Flowing things to/from `AnyT` is generally useless because of the `if not (ground_subtype ...)` short circuit at the top of `__flow`, but this did catch a couple of places where we dropped the whole return type on the floor (`{...(x: Object)}: number` was valid, for instance), and the taint stuff relies on flowing `TaintT ~> AnyT`, which this does without needing `parameters_of_taint_op`.
Closes https://github.com/facebook/flow/pull/1125

Reviewed By: rcastano

Differential Revision: D2690948

Pulled By: mroch

fb-gh-sync-id: 7c538b57499c3ba06faaf9ab74a551e0deaef1c7