fix location of arrow functions with parens
Summary:
fixes the location of `ArrowFunction` ast nodes when the body ends with parens. previously, we'd ignore the parens even though we included the opening parens, and any parens around the arrow function's params.

we use these positions for highlighting errors, for type-at-pos, and eventually for code folding, refactoring or expanding selection, so it's important that it covers the balanced parens.

Reviewed By: gabelevi

Differential Revision: D2715440

fb-gh-sync-id: 15ace0cfd0fcf5d3bfe1880d30c2b7a406060dc2