fix fatal in new error format printer
Summary:
if the error position exceeds the length of the actual line, the new error printer fatals with an `Invalid_argument`.

this makes the printer handle it. there's likely also an underlying bug in that the error position doesn't match with the source.

Fixes https://github.com/facebook/flow/issues/1148

Reviewed By: jeffmo

Differential Revision: D2718572

fb-gh-sync-id: b4aa166357d097ceafac363cda32f802b538ec00