[hack] Reduce number of workers used in Type-check phase for many core machines
Summary: Currently the number of workers we use is determined by the number of CPU cores present on the machine. Each worker has a non-trivial amount of memory overhead. For machines with many cores (right now > 16) we can significantly decrease the amount of memory utilized by the type checker if we use fewer workers with larger buckets.

Reviewed By: int3

Differential Revision: D2719528

fb-gh-sync-id: 2de910d11447c3fac8f664837ac89f899d4ec615